code works! better image iteration
replaced grabImage with operator() on a struct that iterates through
images same as before. This is better because You only need to know
starting dir when you call the constructor and then after that, it
remembers where it is and you don't need any arguments to get the next
image.