code works. can register with MyGL to recieve input callaback
from collage function can register a function with mygl that gets called
with the mouse position every time that input happens. the function gets
added to a vector of functions and mygl runs through all of them calling
each with the x and y pos from glfw. mygl has a function registered with
glfw as a callback function